Release/2.0.0

+++ src/Service/ProjectBillingService.php
@@ -2,7 +2,7 @@
 
 namespace App\Service;
 
-use App\Entity\Account;
+use App\Entity\Client;
 use App\Entity\Invoice;
 use App\Entity\InvoiceEntry;
 use App\Entity\Issue;
@@ -11,43 +11,83 @@ use App\Entity\Worklog;
 use App\Enum\ClientTypeEnum;
 use App\Enum\InvoiceEntryTypeEnum;
 use App\Enum\MaterialNumberEnum;
+use App\Exception\EconomicsException;
 use App\Exception\InvoiceAlreadyOnRecordException;
-use App\Repository\AccountRepository;
 use App\Repository\ClientRepository;
 use App\Repository\IssueRepository;
 use App\Repository\ProjectBillingRepository;
-use App\Repository\WorklogRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 class ProjectBillingService
 {
+    public const PROJECT_BILLING_VERSION_PREFIX = 'PB-';
+
     public function __construct(
-        private readonly AccountRepository $accountRepository,
         private readonly ProjectBillingRepository $projectBillingRepository,
         private readonly BillingService $billingService,
-        private readonly ApiServiceInterface $apiService,
         private readonly IssueRepository $issueRepository,
         private readonly ClientRepository $clientRepository,
-        private readonly WorklogRepository $worklogRepository,
         private readonly EntityManagerInterface $entityManager,
-        private readonly string $receiverAccount,
+        private readonly TranslatorInterface $translator,
+        private readonly string $invoiceDefaultReceiverAccount,
     ) {
     }
 
     /**
+     * @throws EconomicsException
      * @throws \Exception
+     */
+    public function getIssuesNotIncludedInProjectBilling(ProjectBilling $projectBilling): array
+    {
+        $project = $projectBilling->getProject();
+        $periodStart = $projectBilling->getPeriodStart();
+        $periodEnd = $projectBilling->getPeriodEnd();
+
+        if (null == $project) {
+            throw new EconomicsException($this->translator->trans('exception.project_billing_no_project_selected'));
+        }
+
+        if (null == $periodStart || null == $periodEnd) {
+            throw new EconomicsException($this->translator->trans('exception.project_billing_period_cannot_be_null'));
+        }
+
+        $from = new \DateTime($periodStart->format('Y-m-d').' 00:00:00');
+        $to = new \DateTime($periodEnd->format('Y-m-d').' 23:59:59');
+
+        $issues = $this->issueRepository->getClosedIssuesFromInterval($project, $from, $to);
+
+        $filteredIssues = [];
+
+        /** @var Issue $issue */
+        foreach ($issues as $issue) {
+            foreach ($issue->getWorklogs() as $worklog) {
+                if (!$worklog->isBilled() && $worklog->getTimeSpentSeconds() > 0 && null === $worklog->getInvoiceEntry()) {
+                    $filteredIssues[] = $issue;
+                    break;
+                }
+            }
+        }
+
+        return $filteredIssues;
+    }
+
+    /**
+     * @throws EconomicsException
      */
     public function updateProjectBilling(int $projectBillingId): void
     {
         $projectBilling = $this->projectBillingRepository->find($projectBillingId);
 
         if (null == $projectBilling) {
-            // TODO: Replace with custom exception.
-            throw new \Exception('No project billing entity found.');
+            throw new EconomicsException($this->translator->trans('exception.project_billing_no_entity_found'));
         }
 
+        // Delete all non-recorded (ikke-bogført) invoices.
         foreach ($projectBilling->getInvoices() as $invoice) {
-            $this->entityManager->remove($invoice);
+            if (!$invoice->isRecorded()) {
+                $this->entityManager->remove($invoice);
+            }
         }
 
         $this->entityManager->flush();
@@ -56,42 +96,35 @@ class ProjectBillingService
     }
 
     /**
-     * @throws \Exception
+     * @throws EconomicsException
      */
     public function createProjectBilling(int $projectBillingId): void
     {
         $projectBilling = $this->projectBillingRepository->find($projectBillingId);
 
-        // TODO: Replace with custom exception.
         if (null == $projectBilling) {
-            throw new \Exception('No project billing entity found.');
+            throw new EconomicsException($this->translator->trans('exception.project_billing_no_entity_found'));
         }
 
         $project = $projectBilling->getProject();
 
         if (null == $project) {
-            throw new \Exception('No project selected.');
-        }
-
-        $projectTrackerId = $project->getProjectTrackerId();
-
-        if (null == $projectTrackerId) {
-            throw new \Exception('No project.projectTrackerId.');
+            throw new EconomicsException($this->translator->trans('exception.project_billing_no_project_selected'));
         }
 
         $periodStart = $projectBilling->getPeriodStart();
         $periodEnd = $projectBilling->getPeriodEnd();
 
         if (null == $periodStart || null == $periodEnd) {
-            throw new \Exception('project.periodStart or project.periodEnd cannot be null.');
+            throw new EconomicsException($this->translator->trans('exception.project_billing_period_cannot_be_null'));
         }
 
         // Find issues in interval
-        // Group by account
-        // Foreach Account
-        // Create invoice
-        //   Create an InvoiceEntry in the invoice
-        //   Connect worklogs from the database to the invoice entry.
+        // Group by client
+        // Foreach client
+        //   Create invoice
+        //     Create an InvoiceEntry in the invoice
+        //     Connect worklogs from the database to the invoice entry.
         $issues = $this->issueRepository->getClosedIssuesFromInterval($project, $periodStart, $periodEnd);
 
         // TODO: Replace with Model.
@@ -99,56 +132,61 @@ class ProjectBillingService
 
         /** @var Issue $issue */
         foreach ($issues as $issue) {
-            if (null !== $issue->getAccountId()) {
-                $accountId = $issue->getAccountId();
+            $foundProjectBillingVersions = [];
 
-                if (!is_null($accountId)) {
-                    if (!isset($invoices[$accountId])) {
-                        $account = $this->accountRepository->findOneBy(['projectTrackerId' => $issue->getAccountId()]);
-
-                        if (null == $account) {
-                            continue;
-                        }
-
-                        $invoices[$accountId] = [
-                            'account' => $account,
-                            'issues' => [],
-                        ];
-                    }
-
-                    $invoices[$accountId]['issues'][] = $issue;
+            foreach ($issue->getVersions() as $version) {
+                $name = $version->getName();
+                if (null !== $name && str_starts_with($name, self::PROJECT_BILLING_VERSION_PREFIX)) {
+                    $foundProjectBillingVersions[] = $version;
                 }
+            }
+
+            if (1 !== count($foundProjectBillingVersions)) {
+                continue;
+            }
+
+            $foundVersion = $foundProjectBillingVersions[0];
+
+            // Find the client.
+            $client = $this->clientRepository->findOneBy(['versionName' => $foundVersion->getName()]);
+
+            if (null === $client) {
+                continue;
+            }
+
+            $clientId = $client->getId();
+
+            if (null !== $clientId) {
+                if (!isset($invoices[$clientId])) {
+                    $invoices[$clientId] = [
+                        'client' => $client,
+                        'issues' => [],
+                    ];
+                }
+
+                $invoices[$clientId]['issues'][] = $issue;
             }
         }
 
         foreach ($invoices as $invoiceArray) {
-            /** @var Account $account */
-            $account = $invoiceArray['account'];
+            /** @var Client $client */
+            $client = $invoiceArray['client'];
 
             $invoice = new Invoice();
             $invoice->setRecorded(false);
             $invoice->setProject($projectBilling->getProject());
             $invoice->setProjectBilling($projectBilling);
             $invoice->setDescription($projectBilling->getDescription());
-            $invoice->setName($project->getName().': '.$account->getName().' ('.$periodStart->format('d/m/Y').' - '.$periodEnd->format('d/m/Y').')');
+            $invoice->setName($project->getName().': '.$client->getName().' ('.$periodStart->format('d/m/Y').' - '.$periodEnd->format('d/m/Y').')');
             $invoice->setPeriodFrom($periodStart);
             $invoice->setPeriodTo($periodEnd);
-
-            // Find client.
-            $client = $this->clientRepository->findOneBy(['name' => $account->getName(), 'account' => $account->getValue()]);
-
-            // Ignore invoices where there is not client.
-            if (null == $client) {
-                continue;
-            }
-
             $invoice->setClient($client);
 
             $internal = ClientTypeEnum::INTERNAL == $client->getType();
 
             // TODO: MaterialNumberEnum::EXTERNAL_WITH_MOMS or MaterialNumberEnum::EXTERNAL_WITHOUT_MOMS?
             $invoice->setDefaultMaterialNumber($internal ? MaterialNumberEnum::INTERNAL : MaterialNumberEnum::EXTERNAL_WITH_MOMS);
-            $invoice->setDefaultReceiverAccount($this->receiverAccount);
+            $invoice->setDefaultReceiverAccount($this->invoiceDefaultReceiverAccount);
 
             /** @var Issue $issue */
             foreach ($invoiceArray['issues'] as $issue) {
@@ -156,7 +194,7 @@ class ProjectBillingService
                 $invoiceEntry->setEntryType(InvoiceEntryTypeEnum::WORKLOG);
                 $invoiceEntry->setDescription('');
 
-                $product = $issue->getProjectTrackerKey().':'.preg_replace('/\(DEVSUPP-\d+\)/i', '', $issue->getName() ?? '');
+                $product = $this->getInvoiceEntryProduct($issue);
                 $price = $client->getStandardPrice();
 
                 $invoiceEntry->setProduct($product);
@@ -168,7 +206,9 @@ class ProjectBillingService
 
                 /** @var Worklog $worklog */
                 foreach ($worklogs as $worklog) {
-                    if (!$worklog->isBilled()) {
+                    // The invoice entry in the worklog may be non-null, but the entry may have a null id
+                    // @TODO Investigate why this is the case
+                    if (!$worklog->isBilled() && null === $worklog->getInvoiceEntry()?->getId()) {
                         $invoiceEntry->addWorklog($worklog);
                     }
                 }
@@ -221,5 +261,10 @@ class ProjectBillingService
 
         $projectBilling->setRecorded(true);
         $this->projectBillingRepository->save($projectBilling, true);
+    }
+
+    public function getInvoiceEntryProduct(Issue $issue): string
+    {
+        return $issue->getProjectTrackerKey().':'.preg_replace('/\(DEVSUPP-\d+\)/i', '', $issue->getName() ?? '');
     }
 }
